feat(simulacrum): move `TransactionExecutor` to `simulacrum`
