+++ crates/iota-analytics-indexer/src/handlers/transaction_handler.rs
@@ -212,7 +212,7 @@ mod tests {
 
     #[tokio::test]
     pub async fn test_transaction_handler() -> anyhow::Result<()> {
-        let mut sim = Simulacrum::new();
+        let sim = Simulacrum::new();
 
         // Execute a simple transaction.
         let transfer_recipient = IotaAddress::random_for_testing_only();
